+++ pages/_app.tsx
@@ -10,8 +10,8 @@ import theme from '../styles/initAppTheme';
 import { createEmotionCache } from '@/lib/emotion';
 import { SWRConfig } from 'swr';
 import { fetcher, axios } from '../lib/axios';
-import { AuthProvider } from '@/services/context/AuthProvider';
-import { RouteGuardProvider } from '@/services/context/RouteGuardProvider';
+import { AuthProvider } from '@/providers/AuthProvider';
+import { RouteGuardProvider } from '@/providers/RouteGuardProvider';
 import Backdrop from '@mui/material/Backdrop';
 import CircularProgress from '@mui/material/CircularProgress';
 import { SessionProvider } from 'next-auth/react';
