+++ providers/AuthProvider.tsx
@@ -1,7 +1,7 @@
 import React, { createContext, ReactNode, useRef } from 'react';
 import { LoginPost } from '../schemas/LoginSchema';
 import useSWR from 'swr';
-import { fetcher, setToken } from '../../lib/axios';
+import { fetcher, setToken } from '@/lib/axios';
 import { User } from 'lib/swagger';
 import { swaggerClient } from 'lib/swaggerClient';
 import { UserType } from '../types/types';
